style: 리스트 페이지 퍼블리싱 완료
