--- src/app/(global)/keyboards_back/page.tsx
@@ -1,12 +1,0 @@
-import KeyboardsArea from '@/components/feature/Keyboards/KeyboardsArea';
-
-const Page = () => {
-  return (
-    <div className='px-4 pt-[15px] pb-[100px] m-auto max-w-[1140px] md:pt-5 md:pb-[50px] md:px-5 lg:px-0'>
-      <div>슬라이드바 영역</div>
-      <KeyboardsArea />
-    </div>
-  );
-};
-
-export default Page;
